pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "options-wheel"
version = "0.1.0"
description = "An automated options wheel trading strategy."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "pandas>=1.5",
    "numpy>=1.23",
    "requests>=2.28",
    "alpaca-py"
]

[project.scripts]
run-strategy = "scripts.run_strategy:main"
# (optional) lets users just type `run-strategy` in the terminal

[tool.setuptools.packages.find]
where = ["."]
exclude = ["archive"]

